refactor: move metadata-relevant methods into an independent socket file
