+++ backend/webserver/sockets/document.js
@@ -840,522 +840,6 @@ class DocumentSocket extends Socket {
     }
 
     /**
-     * Ensure the caller may import metadata for the given assignment.
-     *
-     * Access is granted when the assignment is open and the caller is any of:
-     * - a global admin,
-     * - the assignment owner,
-     * - a user with `frontend.dashboard.assignments.edit` (same right used for editing assignments in the dashboard).
-     *
-     * @param {Object} assignment
-     * @param {Object} [options={}]
-     */
-    async assertMetadataImportAccess(assignment, options = {}) {
-        if (assignment.closed) {
-            throw new Error("Cannot import metadata because the assignment is closed.");
-        }
-        if (await this.isAdmin()) return;
-
-        if (Number(assignment.userId) === Number(this.userId)) return;
-
-        if (await this.hasAccess("frontend.dashboard.assignments.edit")) return;
-
-        throw new Error("You do not have permission to import metadata for this assignment.");
-    }
-
-    /**
-     * Persist mapped metadata for all documents of one matched submission row.
-     *
-     * @param {Object} data
-     * @param {number[]} data.documentIds
-     * @param {number} data.userId
-     * @param {Object} data.row
-     * @param {Object[]} data.mappings
-     * @param {string|null} data.fileName
-     * @param {Object} options
-     * @returns {Promise<number>}
-     */
-    async attachMappedMetadataToDocuments(data, options = {}) {
-        const entries = [];
-        const fileName = data.fileName || "";
-
-        for (const documentId of data.documentIds) {
-            for (const mapping of data.mappings) {
-                if (!Object.prototype.hasOwnProperty.call(data.row, mapping.sourceField)) {
-                    continue;
-                }
-
-                const metaValue = data.row[mapping.sourceField] == null ? "" : String(data.row[mapping.sourceField]);
-                entries.push(
-                    {documentId, userId: data.userId, metaKey: mapping.metaKey, metaValue},
-                    {documentId, userId: data.userId, metaKey: `${mapping.metaKey}.sourceFile`, metaValue: fileName},
-                    {documentId, userId: data.userId, metaKey: `${mapping.metaKey}.sourceField`, metaValue: mapping.sourceField},
-                );
-            }
-        }
-
-        if (entries.length === 0) {
-            return 0;
-        }
-
-        await this.models["document_metadata"].bulkUpsertByDocumentAndKey(entries, options);
-        return entries.length;
-    }
-
-    /**
-     * Normalize metadata mappings into a canonical backend shape.
-     *
-     * @param {Object[]} mappings
-     * @returns {Object[]}
-     */
-    normalizeMetadataMappings(mappings = []) {
-        return mappings
-            .map((mapping) => ({
-                sourceField: String(mapping?.sourceField || "").trim(),
-                metaKey: String(mapping?.metaKey || "").trim(),
-            }))
-            .filter((mapping) => mapping.sourceField && mapping.metaKey);
-    }
-
-    /**
-     * Validate target metadata keys before adding system-generated provenance rows.
-     *
-     * @param {Object[]} mappings
-     */
-    validateMetadataMappings(mappings = []) {
-        const reservedSuffixes = [".sourceFile", ".sourceField"];
-        const metaKeys = mappings.map((mapping) => mapping.metaKey);
-
-        if (new Set(metaKeys).size !== metaKeys.length) {
-            throw new Error("Target metaKeys must be unique.");
-        }
-
-        const reservedMetaKeys = metaKeys.filter((metaKey) => (
-            reservedSuffixes.some((suffix) => metaKey.endsWith(suffix))
-        ));
-        if (reservedMetaKeys.length > 0) {
-            throw new Error(
-                `Target metaKeys cannot end with reserved provenance suffixes: ${reservedMetaKeys.join(", ")}`
-            );
-        }
-    }
-
-    /**
-     * Normalize the primary key mapping into a canonical backend shape.
-     *
-     * @param {Object} primaryKeyMapping
-     * @returns {Object}
-     */
-    normalizePrimaryKeyMapping(primaryKeyMapping = {}) {
-        return {
-            sourceField: String(primaryKeyMapping?.sourceField || "").trim(),
-            targetField: String(primaryKeyMapping?.targetField || "").trim(),
-        };
-    }
-
-    /**
-     * Normalize one primary key value according to the selected target field.
-     *
-     * @param {*} rawValue
-     * @param {string} targetField
-     * @returns {number|string|null}
-     */
-    normalizePrimaryKeyValue(rawValue, targetField) {
-        if (rawValue == null) {
-            return null;
-        }
-
-        const stringValue = String(rawValue).trim();
-        if (!stringValue) {
-            return null;
-        }
-
-        if (targetField === "extId") {
-            const numericValue = Number(stringValue);
-            return Number.isNaN(numericValue) ? null : numericValue;
-        }
-
-        if (targetField === "email") {
-            return stringValue.toLowerCase();
-        }
-
-        return null;
-    }
-
-    /**
-     * Validate primary key values for emptiness and duplicates.
-     *
-     * @param {Object[]} rows
-     * @param {Object} primaryKeyMapping
-     */
-    validatePrimaryKeyValues(rows, primaryKeyMapping) {
-        const seen = new Set();
-        const duplicates = new Set();
-
-        for (const row of rows) {
-            const normalized = this.normalizePrimaryKeyValue(row?.[primaryKeyMapping.sourceField], primaryKeyMapping.targetField);
-            if (normalized == null) {
-                throw new Error("Primary key values must be present and valid for every imported row.");
-            }
-
-            if (seen.has(normalized)) {
-                duplicates.add(normalized);
-            } else {
-                seen.add(normalized);
-            }
-        }
-
-        if (duplicates.size > 0) {
-            throw new Error(
-                `Duplicate primary key values found for ${primaryKeyMapping.targetField}: ${[...duplicates].join(", ")}`
-            );
-        }
-    }
-
-    /**
-     * Resolve a metadata import row against assignment submissions.
-     *
-     * @param {*} rawValue
-     * @param {string} targetField
-     * @param {Map<number, Object[]>} submissionByExtId
-     * @param {Map<string, Object[]>} submissionByEmail
-     * @returns {Object[]}
-     */
-    resolveMetadataImportSubmission(rawValue, targetField, submissionByExtId, submissionByEmail) {
-        const normalizedValue = this.normalizePrimaryKeyValue(rawValue, targetField);
-        if (normalizedValue == null) {
-            return [];
-        }
-
-        if (targetField === "extId" && submissionByExtId.has(normalizedValue)) {
-            return submissionByExtId.get(normalizedValue);
-        }
-
-        if (targetField === "email" && submissionByEmail.has(normalizedValue)) {
-            return submissionByEmail.get(normalizedValue);
-        }
-
-        return [];
-    }
-
-    /**
-     * Build a metadata import plan for one assignment target.
-     *
-     * Validates caller access, normalizes mappings, and resolves uploaded rows to
-     * submission documents without writing metadata.
-     *
-     * @param {Object} data
-     * @param {string} [data.targetType="assignment"] - Import target type; only `"assignment"` is supported.
-     * @param {number} data.assignmentId - Assignment whose submission documents receive metadata.
-     * @param {Object} data.primaryKeyMapping - How uploaded rows are matched to submission owners.
-     * @param {string} data.primaryKeyMapping.sourceField - Column/key from each uploaded row.
-     * @param {"extId"|"email"} data.primaryKeyMapping.targetField - Submission-owner field to match against.
-     * @param {Object[]} data.mappings - Field mappings from uploaded rows to document `metaKey` values.
-     * @param {string} data.mappings[].sourceField - Source column/key in each uploaded row.
-     * @param {string} data.mappings[].metaKey - Target metadata key written on matched documents.
-     * @param {Object[]} data.rows - Parsed metadata rows from the uploaded file.
-     * @param {string} [data.fileName] - Original upload filename stored as metadata provenance.
-     * @param {Object} [options={}]
-     * @param {Object} [options.transaction] - Sequelize transaction passed to DB reads.
-     * @returns {Promise<Object>} Import plan with counts plus `matched`, `unmatched`, `skipped`, and `overwritten` details.
-     */
-    async buildMetadataImportPlan(data, options = {}) {
-        const targetType = data.targetType || "assignment";
-        if (targetType !== "assignment") {
-            throw new Error(`Unsupported target type "${targetType}".`);
-        }
-
-        const assignmentId = Number(data.assignmentId || 0);
-        if (!assignmentId) {
-            throw new Error("Assignment ID is required.");
-        }
-
-        const assignment = await this.models["assignment"].getById(assignmentId, {transaction: options.transaction});
-        if (!assignment) {
-            throw new Error(`Assignment with id ${assignmentId} not found`);
-        }
-
-        // Check if caller has the right to import metadata
-        await this.assertMetadataImportAccess(assignment, options);
-
-        const primaryKeyMapping = this.normalizePrimaryKeyMapping(data.primaryKeyMapping);
-        if (!primaryKeyMapping.sourceField) {
-            throw new Error("Primary key source field is required.");
-        }
-        if (!["extId", "email"].includes(primaryKeyMapping.targetField)) {
-            throw new Error("Primary key target field must be either \"extId\" or \"email\".");
-        }
-
-        const rows = Array.isArray(data.rows) ? data.rows.filter((row) => row && typeof row === "object") : [];
-        if (rows.length === 0) {
-            throw new Error("No metadata rows provided.");
-        }
-        this.validatePrimaryKeyValues(rows, primaryKeyMapping);
-
-        const mappings = this.normalizeMetadataMappings(data.mappings);
-        if (mappings.length === 0) {
-            throw new Error("At least one metadata mapping is required.");
-        }
-        this.validateMetadataMappings(mappings);
-
-        const submissions = await this.models["submission"].findAll({
-            where: {
-                assignmentId,
-                deleted: false,
-            },
-            raw: true,
-            transaction: options.transaction,
-        });
-
-        const ownerUserIds = [...new Set(submissions.map((submission) => Number(submission.userId)))];
-        const users = ownerUserIds.length > 0
-            ? await this.models["user"].findAll({
-                where: {
-                    id: ownerUserIds,
-                    deleted: false,
-                },
-                attributes: ["id", "extId", "email"],
-                raw: true,
-                transaction: options.transaction,
-            })
-            : [];
-        const usersById = new Map(users.map((user) => [Number(user.id), user]));
-        const submissionByExtId = new Map();
-        const submissionByEmail = new Map();
-
-        for (const submission of submissions) {
-            const owner = usersById.get(Number(submission.userId));
-            if (!owner) {
-                continue;
-            }
-
-            if (owner.extId != null) {
-                const extId = Number(owner.extId);
-                if (!submissionByExtId.has(extId)) {
-                    submissionByExtId.set(extId, []);
-                }
-                submissionByExtId.get(extId).push(submission);
-            }
-
-            if (owner.email) {
-                const email = String(owner.email).trim().toLowerCase();
-                if (!submissionByEmail.has(email)) {
-                    submissionByEmail.set(email, []);
-                }
-                submissionByEmail.get(email).push(submission);
-            }
-        }
-
-        const documents = submissions.length > 0
-            ? await this.models["document"].findAll({
-                where: {
-                    submissionId: submissions.map((submission) => submission.id),
-                    deleted: false,
-                },
-                raw: true,
-                transaction: options.transaction,
-            })
-            : [];
-
-        const documentsBySubmissionId = documents.reduce((acc, document) => {
-            if (!acc.has(document.submissionId)) {
-                acc.set(document.submissionId, []);
-            }
-            acc.get(document.submissionId).push(document);
-            return acc;
-        }, new Map());
-
-        const metadataKeys = mappings.flatMap((mapping) => [
-            mapping.metaKey,
-            `${mapping.metaKey}.sourceFile`,
-            `${mapping.metaKey}.sourceField`,
-        ]);
-
-        const existingMetadataRows = documents.length > 0
-            ? await this.models["document_metadata"].findAll({
-                where: {
-                    documentId: documents.map((document) => document.id),
-                    metaKey: metadataKeys,
-                    deleted: false,
-                },
-                raw: true,
-                transaction: options.transaction,
-            })
-            : [];
-
-        const existingByDocumentId = existingMetadataRows.reduce((acc, row) => {
-            const documentId = Number(row.documentId);
-            if (!acc.has(documentId)) {
-                acc.set(documentId, new Set());
-            }
-            acc.get(documentId).add(row.metaKey);
-            return acc;
-        }, new Map());
-
-        const matched = [];
-        const unmatched = [];
-        const overwritten = [];
-        const skipped = [];
-
-        let documentCount = 0;
-        let metadataEntryCount = 0;
-        let overwrittenEntryCount = 0;
-
-        for (const row of rows) {
-            const primaryKeyValue = row[primaryKeyMapping.sourceField];
-            const resolvedSubmissions = this.resolveMetadataImportSubmission(
-                primaryKeyValue,
-                primaryKeyMapping.targetField,
-                submissionByExtId,
-                submissionByEmail
-            );
-
-            if (resolvedSubmissions.length === 0) {
-                unmatched.push({
-                    primaryKeyValue: primaryKeyValue ?? null,
-                    message: "No submission owner match found in this assignment.",
-                });
-                continue;
-            }
-
-            const ownerUserIdsForRow = new Set(resolvedSubmissions.map((submission) => Number(submission.userId)));
-            if (ownerUserIdsForRow.size > 1) {
-                unmatched.push({
-                    primaryKeyValue: primaryKeyValue ?? null,
-                    message: "Primary key matched submissions owned by different users.",
-                });
-                continue;
-            }
-
-            const effectiveMappings = mappings.filter((mapping) => Object.prototype.hasOwnProperty.call(row, mapping.sourceField));
-            if (effectiveMappings.length === 0) {
-                skipped.push({
-                    submissionId: resolvedSubmissions[0]?.id || null,
-                    message: "Row does not contain any mapped source fields.",
-                });
-                continue;
-            }
-
-            const submissionDocuments = [];
-            for (const submission of resolvedSubmissions) {
-                const documentsForSubmission = documentsBySubmissionId.get(submission.id) || [];
-                if (documentsForSubmission.length === 0) {
-                    skipped.push({
-                        submissionId: submission.id,
-                        message: "Submission has no documents.",
-                    });
-                    continue;
-                }
-                submissionDocuments.push(...documentsForSubmission);
-            }
-
-            if (submissionDocuments.length === 0) {
-                continue;
-            }
-
-            let rowOverwrittenEntryCount = 0;
-
-            for (const document of submissionDocuments) {
-                const existingKeys = existingByDocumentId.get(Number(document.id)) || new Set();
-                for (const mapping of effectiveMappings) {
-                    for (const metaKey of [mapping.metaKey, `${mapping.metaKey}.sourceFile`, `${mapping.metaKey}.sourceField`]) {
-                        if (existingKeys.has(metaKey)) {
-                            rowOverwrittenEntryCount += 1;
-                        }
-                    }
-                }
-            }
-
-            const rowMetadataEntryCount = submissionDocuments.length * effectiveMappings.length * 3;
-            matched.push({
-                userId: resolvedSubmissions[0].userId,
-                row,
-                documentIds: submissionDocuments.map((document) => document.id),
-                mappings: effectiveMappings,
-            });
-
-            if (rowOverwrittenEntryCount > 0) {
-                overwritten.push({
-                    submissionId: resolvedSubmissions[0].id,
-                    message: `${rowOverwrittenEntryCount} metadata entries were overwritten.`,
-                });
-            }
-
-            documentCount += submissionDocuments.length;
-            metadataEntryCount += rowMetadataEntryCount;
-            overwrittenEntryCount += rowOverwrittenEntryCount;
-        }
-
-        return {
-            matchedRowCount: matched.length,
-            unmatchedRowCount: unmatched.length,
-            skippedRowCount: skipped.length,
-            documentCount,
-            metadataEntryCount,
-            overwrittenEntryCount,
-            matched,
-            unmatched,
-            overwritten,
-            skipped,
-        };
-    }
-
-    /**
-     * Preview metadata import without mutating documents.
-     *
-     * Authorization is enforced inside `buildMetadataImportPlan`.
-     *
-     * @param {Object} data - Same payload as `buildMetadataImportPlan`.
-     * @param {Object} [options={}]
-     * @returns {Promise<Object>} Summary counts only (no `matched` / `unmatched` row details).
-     */
-    async previewMetadataImport(data, options = {}) {
-        const plan = await this.buildMetadataImportPlan(data, options);
-        return {
-            matchedRowCount: plan.matchedRowCount,
-            unmatchedRowCount: plan.unmatchedRowCount,
-            skippedRowCount: plan.skippedRowCount,
-            documentCount: plan.documentCount,
-            metadataEntryCount: plan.metadataEntryCount,
-            overwrittenEntryCount: plan.overwrittenEntryCount,
-        };
-    }
-
-    /**
-     * Import metadata rows for existing submissions in one assignment.
-     *
-     * Authorization is enforced inside `buildMetadataImportPlan`.
-     *
-     * @param {Object} data - Same payload as `buildMetadataImportPlan`.
-     * @param {Object} [options={}]
-     * @returns {Promise<Object>} Import summary counts plus unmatched, skipped, and overwritten row details.
-     */
-    async importMetadata(data, options = {}) {
-        const plan = await this.buildMetadataImportPlan(data, options);
-
-        for (const match of plan.matched) {
-            await this.attachMappedMetadataToDocuments({
-                documentIds: match.documentIds,
-                userId: match.userId,
-                row: match.row,
-                mappings: match.mappings,
-                fileName: data.fileName || null,
-            }, options);
-        }
-
-        return {
-            matchedRowCount: plan.matchedRowCount,
-            unmatchedRowCount: plan.unmatchedRowCount,
-            skippedRowCount: plan.skippedRowCount,
-            documentCount: plan.documentCount,
-            metadataEntryCount: plan.metadataEntryCount,
-            unmatched: plan.unmatched,
-            skipped: plan.skipped,
-            overwritten: plan.overwritten,
-        };
-    }
-
-    /**
      * Downloads multiple submission files from Moodle URLs, creating a local document record for each one.
      * Each file is processed in its own database transaction to ensure atomicity. Progress is reported
      * to the client via a socket event after each file is processed.
@@ -1472,7 +956,7 @@ class DocumentSocket extends Socket {
                 if (topicAllocation) {
                     try {
                         const topicValue = topicAllocation.topicName ?? topicAllocation.topic ?? null;
-                        await this.attachMappedMetadataToDocuments({
+                        await this.getSocket("DocumentMetadataSocket").attachMappedMetadataToDocuments({
                             documentIds,
                             userId: submission.userId,
                             row: { topic: topicValue },
@@ -2191,8 +1675,6 @@ class DocumentSocket extends Socket {
         this.createSocket("documentOpen", this.openDocument, {}, false);
         this.createSocket("documentGetAll", this.refreshAllDocuments, {}, false);
         this.createSocket("documentUploadSingleSubmission", this.uploadSingleSubmission, {}, true);
-        this.createSocket("documentPreviewMetadataImport", this.previewMetadataImport, {}, false);
-        this.createSocket("documentImportMetadata", this.importMetadata, {}, true);
     }
 };
